requirements.txt
requests==2.32.4,
pydantic==2.11.7,
python-dotenv==1.1.1,